add : jpa2 code add
