+++ __Spring_Master/HGH/Code/SpringBoot-실전! 스프링 부트와 JPA활용2/jpashop/src/main/java/jpabook/jpashop/service/MemberService.java
@@ -50,7 +50,7 @@ public class MemberService {
     }
 
     public Member findOne(Long memberId){
-        return memberRepository.findOne(memberId);
+        return memberRepository.findById(memberId).get();
     }
